// src/app/models/salary-details.model.ts
export class SalaryDetails {
    ctc: number;
    basic: number;
    annualBasic: number;
    calcOnMinBasic: boolean;
    epfEmployer: number;
    epfEmployee: number;
    inHand: number;
    tax: number;
    monthlyEpfEmployer: number;
    monthlyEpfEmployee: number;
    hra: number;
    metroResident: boolean;
    rentPaid: number;
    hraMonthly: number;
    hraExemption: number;
    invest80C: number;
    monthlyTax: number;
    gratuity: number;
    maximize80C: boolean;
    foodCouponsAvailable: boolean;
    foodCouponsOpted: boolean;
    foodCouponsDeduction: number;
    carLeaseAvailable: boolean;
    optedForMediclaim: boolean;
    annualMediclaimAmount: number;
    monthlyMediclaimAmount: number;
    monthlyCarLeaseAmount: number;
    monthlyDriversSalaryAmount: number;
    monthlyFuelAndMaintainenceAmount: number;
    yearlyCarLeaseAmount: number;
    yearlyDriversSalaryAmount: number;
    yearlyFuelAndMaintainenceAmount: number;
    netTaxableIncome: number;
}